refactor: move soundcloud_tools into backend
What survived the dead-code removal was no longer a SoundCloud library:
it is local audio-file metadata handling plus one OAuth helper. Renamed
and relocated accordingly, and the handler/ + utils/ split was dropped
since only three modules remained.

    handler/track.py       -> backend/core/audio/tags.py
    handler/folder.py      -> backend/core/audio/folders.py
    utils/string.py     +
    handler/artist_ranking -> backend/core/audio/titles.py
    oauth.py               -> backend/core/services/sc_oauth.py
    settings.py            -> backend/sc_settings.py

sc_oauth.py now sits beside sc_auth_cache.py, its only consumer.
sc_settings.py stays a separate BaseSettings rather than merging into
BackendSettings, which declares env_prefix="BACKEND_" while these read
unprefixed CLIENT_ID / OAUTH_TOKEN / USER_ID / ROOT_MUSIC_FOLDER.

FILETYPE_MAP moved from tags.py to folders.py to break a circular import
(tags needed load_tracks, folders needed FILETYPE_MAP). Dependencies now
run one way: tags -> folders. convert_to_int moved into tags.py, its only
caller.
