+++ backend/core/services/collection.py
@@ -14,10 +14,9 @@ from pathlib import Path
 
 from pydantic import ValidationError
 
+from backend.core.audio.folders import FolderHandler, load_tracks, load_tracks_recursive
+from backend.core.audio.tags import TrackHandler, TrackInfo
 from backend.core.services import cache_db
-from soundcloud_tools.handler.folder import FolderHandler
-from soundcloud_tools.handler.track import TrackHandler, TrackInfo
-from soundcloud_tools.utils import load_tracks, load_tracks_recursive
 
 logger = logging.getLogger(__name__)
