+++ scripts/generate_sample_tracks.py
@@ -30,7 +30,7 @@ from dataclasses import dataclass
 from datetime import date, timedelta
 from pathlib import Path
 
-from soundcloud_tools.handler.track import StarlibMeta, TrackHandler, TrackInfo
+from backend.core.audio.tags import StarlibMeta, TrackHandler, TrackInfo
 
 logger = logging.getLogger(__name__)
